pyproject.toml
[tool.poetry]
name = "textension"
version = "0.1.0"
description = ""
authors = ["Nathan Beals <nathan.beals@ontariotechu.net>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
numpy = "^1.19.1"
PyPDF2 = "^1.26.0"
Pillow = "^7.2.0"